requirements/base.txt
chevron==0.13.1
namedlist==1.7
pyyaml==3.13
